api/requirements.txt
uvicorn[standard]
typing
fastapi[standard]
pyenchant